requirements.txt
flask==3.0.0
python-dotenv==1.0.1
youtube_transcript_api>=1.1.0
beautifulsoup4==4.12.3
markdown==3.5.1
litellm>=1.40.0
google-api-python-client==2.116.0
supabase>=2.0.0
gunicorn>=21.0.0
cryptography>=41.0.0
